refactor(pdf): consolidate getPdfJob resolver into orchestrator Lambda
Stack is at 508 resources, CFN hard cap is 500. Moving the getPdfJob
query resolver from its own Lambda into the orchestrator Lambda (same
container image, dispatch via AppSync Powertools resolver on fieldName)
saves ~5-6 CFN resources. Matches DREM's uploadModelToCarAppSyncHandler
pattern of one AppSync handler serving multiple fields.
