+++ lib/lambdas/pdf_api/index.py
@@ -23,9 +23,11 @@ app = AppSyncResolver()
 
 PDF_JOBS_TABLE = os.environ["PDF_JOBS_TABLE"]
 WORKER_FUNCTION_NAME = os.environ["WORKER_FUNCTION_NAME"]
+URL_EXPIRY_SECONDS = int(os.environ.get("URL_EXPIRY_SECONDS", "3600"))
 
 _dynamodb = boto3.resource("dynamodb")
 _lambda = boto3.client("lambda")
+_s3 = boto3.client("s3")
 _jobs_table = _dynamodb.Table(PDF_JOBS_TABLE)
 
 
@@ -64,3 +66,30 @@ def generate_race_results_pdf(eventId: str, type: str, userId: str = None, track
         Payload=json.dumps({"jobId": job_id}),
     )
     return {**item, "downloadUrl": None, "error": None, "completedAt": None, "filename": None}
+
+
+@app.resolver(type_name="Query", field_name="getPdfJob")
+def get_pdf_job(jobId: str):
+    requester = shared.requester_identity((app.current_event.identity or {}))
+    resp = _jobs_table.get_item(Key={"jobId": jobId})
+    item = resp.get("Item")
+    if not item:
+        return None
+    if requester["sub"] != item.get("createdBy") and not (requester["groups"] & shared.ADMIN_GROUPS):
+        raise PermissionError("Not your PDF job")
+
+    result = shared.replace_decimal_with_float(item)
+    result.setdefault("downloadUrl", None)
+    result.setdefault("error", None)
+    result.setdefault("completedAt", None)
+    if item.get("status") == "SUCCESS" and item.get("s3Key") and item.get("filename"):
+        result["downloadUrl"] = _s3.generate_presigned_url(
+            "get_object",
+            Params={
+                "Bucket": shared.PDF_BUCKET,
+                "Key": item["s3Key"],
+                "ResponseContentDisposition": f'attachment; filename="{item["filename"]}"',
+            },
+            ExpiresIn=URL_EXPIRY_SECONDS,
+        )
+    return result
